Move font-size scale to constants layer
Copilot review: `constants/GlobalStyles.ts` and `helpers/utils/color.ts`
imported the scale from `components/typography/`, inverting the dependency
direction (constants/helpers → components) and risking circular deps.

The scale is pure data with no component dependencies, so relocate it to
`src/constants/FontSizes.ts` (via git mv, preserving history) and update all
importers. constants→constants and helpers/components→constants are all
correct-direction dependencies.
