+++ src/components/audio/AudioPlayer.tsx
@@ -8,10 +8,10 @@ import { useEffect, useRef, useState } from "react";
 import { View } from "react-native";
 
 import { PauseIcon, UnmuteIcon } from "#/components/Icons";
-import { fontSizes } from "#/components/typography/fontSizes";
 import UiPressable from "#/components/ui/UiPressable";
 import UiText from "#/components/ui/UiText";
 import Colors from "#/constants/Colors";
+import { fontSizes } from "#/constants/FontSizes";
 import {
   useAppColorScheme,
   useCorporateColor,
